memory_system/requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.9
beautifulsoup4>=4.12.0
